refactor(cda-main): Move com_param code into one module.
The code was spread between `configfile.rs` and `lib.rs`,
with particularly the unit tests largely cross-referencing
between the files.

The test case in `mdd.rs` was also removed, because it was an
exact duplicate of the test case in `lib.rs`.
